requirements.txt
Flask==0.9
Jinja2==2.7
MarkupSafe==0.18
Pillow==2.0.0
Werkzeug==0.8.3
distribute==0.6.31
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
gunicorn==0.16.1
wsgiref==0.1.2
